Ignore unnecessary shards

+++ distributed/shuffle/_shuffle.py
@@ -333,8 +333,10 @@ def split_by_worker(
     out: defaultdict[str, list[tuple[int, list[PickleBuffer]]]] = defaultdict(list)
 
     base = df[column].values.base
-    for output_part_id, part in df.groupby(column, observed=False):
+    for output_part_id, part in df.groupby(column, observed=True):
         assert isinstance(output_part_id, int)
+        if output_part_id not in worker_for:
+            continue
         if part[column].values.base is base and len(part) != len(base):
             if drop_column:
                 del part[column]
